requirements.txt
# Superset is not Pipfile compatible, having dependencies that cannot be safely resolved
pandas==0.23.4
sqlalchemy==1.2.18

superset==0.28.1

pandas==0.23.4
flask-appbuilder==1.12.1
sqlalchemy==1.2.18
sqlalchemy-utils==0.33.11
oauthlib==2.1.0

psycopg2-binary==2.7.7
redis==3.2.1
flask_oauthlib==0.9.5
oauthlib==2.1.0
sentry_sdk
circus
